make pl_gradcache subclass of gradcache
